requirements.txt
python-telegram-bot==20.7
pyautogui==0.9.54
PyGithub==2.1.1
python-dotenv==1.0.0
pyperclip==1.8.2